Fix topology solver public API
