+++ tt_metal/CMakeLists.txt
@@ -92,6 +92,8 @@ set(TT_METAL_PUBLIC_API
     api/tt-metalium/queue_id.hpp
     api/tt-metalium/experimental/fabric/routing_table_generator.hpp
     api/tt-metalium/experimental/fabric/topology_mapper_utils.hpp
+    api/tt-metalium/experimental/fabric/topology_solver.hpp
+    api/tt-metalium/experimental/fabric/topology_solver.tpp
     api/tt-metalium/runtime_args_data.hpp
     api/tt-metalium/shape.hpp
     api/tt-metalium/shape2d.hpp
