+++ tt_metal/api/tt-metalium/experimental/fabric/topology_solver.tpp
@@ -16,7 +16,6 @@
 #include <fmt/format.h>
 #include <tt-logger/tt-logger.hpp>
 #include <tt_stl/assert.hpp>
-#include "tt_metal/fabric/topology_solver_internal.hpp"
 
 namespace tt::tt_fabric {
