+++ tt_metal/fabric/topology_solver_internal.hpp
@@ -27,96 +27,6 @@ struct MappingResult;
 namespace tt::tt_fabric::detail {
 
 /**
- * @brief Indexed graph representation for efficient lookups
- *
- * Converts AdjacencyGraph into index-based representation for O(1) lookups.
- * Stores deduplicated, sorted adjacency lists and connection counts.
- */
-template <typename TargetNode, typename GlobalNode>
-struct GraphIndexData {
-    // Node vectors
-    std::vector<TargetNode> target_nodes;
-    std::vector<GlobalNode> global_nodes;
-
-    // Index mappings
-    std::map<TargetNode, size_t> target_to_idx;
-    std::map<GlobalNode, size_t> global_to_idx;
-
-    // Adjacency index vectors (deduplicated, sorted)
-    std::vector<std::vector<size_t>> target_adj_idx;
-    std::vector<std::vector<size_t>> global_adj_idx;
-
-    // Connection count maps (for strict mode / multi-edge support)
-    // conn_count[i][j] = number of channels from node i to node j
-    std::vector<std::map<size_t, size_t>> target_conn_count;
-    std::vector<std::map<size_t, size_t>> global_conn_count;
-
-    // Degree vectors
-    std::vector<size_t> target_deg;
-    std::vector<size_t> global_deg;
-
-    size_t n_target = 0;
-    size_t n_global = 0;
-
-    /**
-     * @brief Construct GraphIndexData from AdjacencyGraph inputs
-     *
-     * Builds indexed representation from target and global graphs.
-     */
-    GraphIndexData(
-        const AdjacencyGraph<TargetNode>& target_graph, const AdjacencyGraph<GlobalNode>& global_graph);
-
-    /**
-     * @brief Print node degrees for debugging
-     *
-     * Prints the degree of each node in both target and global graphs.
-     * Useful for understanding graph structure during mapping.
-     */
-    void print_node_degrees() const;
-
-    /**
-     * @brief Print adjacency maps for debugging
-     *
-     * Prints the adjacency structure of both target and global graphs.
-     * Useful for debugging mapping failures.
-     */
-    void print_adjacency_maps() const;
-};
-
-/**
- * @brief Indexed constraint representation for efficient lookups
- *
- * Converts MappingConstraints into index-based representation for O(1) constraint checks.
- * Stores restricted and preferred mappings as index vectors.
- */
-template <typename TargetNode, typename GlobalNode>
-struct ConstraintIndexData {
-    // Restricted mappings: target_idx -> vector of valid global_indices
-    // If empty for a target_idx, that target can map to any global node
-    std::vector<std::vector<size_t>> restricted_global_indices;
-
-    // Preferred mappings: target_idx -> vector of preferred global_indices
-    // Used for optimization, doesn't restrict valid mappings
-    std::vector<std::vector<size_t>> preferred_global_indices;
-
-    /**
-     * @brief Construct ConstraintIndexData from MappingConstraints and GraphIndexData
-     *
-     * Builds indexed constraint representation from constraints and graph data.
-     */
-    ConstraintIndexData(
-        const MappingConstraints<TargetNode, GlobalNode>& constraints,
-        const GraphIndexData<TargetNode, GlobalNode>& graph_data);
-
-    // Helper: check if mapping is valid
-    bool is_valid_mapping(size_t target_idx, size_t global_idx) const;
-
-    // Helper: get candidates for target node
-    // Returns restricted candidates if available, otherwise returns empty vector (meaning all are valid)
-    const std::vector<size_t>& get_candidates(size_t target_idx) const;
-};
-
-/**
  * @brief Unified heuristic for node selection and candidate generation
  *
  * Combines node selection and candidate generation with explicit priority:
@@ -304,173 +214,6 @@ struct ConsistencyChecker {
 
 template <typename TargetNode, typename GlobalNode>
 struct PathGraphDetector;
-
-/**
- * @brief DFS search engine for topology mapping
- *
- * Implements backtracking search with memoization and consistency checking.
- * Uses SearchHeuristic for node selection and candidate generation.
- *
- * **Important**: Even if the search fails to find a complete valid mapping, the
- * `SearchState::mapping` will contain the best/closest partial mapping found.
- * This allows users to see what progress was made and diagnose why the search failed.
- * The MappingValidator will save this partial mapping in the result even if validation fails.
- */
-template <typename TargetNode, typename GlobalNode>
-class DFSSearchEngine {
-public:
-    /**
-     * @brief Search state tracking mapping progress and statistics
-     *
-     * **Note**: The `mapping` vector always contains the best mapping found so far,
-     * even if the search fails. This allows users to inspect partial mappings for debugging.
-     */
-    struct SearchState {
-        std::vector<int> mapping;                    // mapping[target_idx] = global_idx or -1 (best found so far)
-        std::vector<bool> used;                      // used[global_idx] = true if assigned
-        std::unordered_set<uint64_t> failed_states;  // Memoization cache of failed states
-        size_t dfs_calls = 0;                        // Number of DFS calls made
-        size_t backtrack_count = 0;                  // Number of backtracks performed
-        std::string error_message;                   // Error message if search fails
-    };
-
-    /**
-     * @brief Start DFS search
-     *
-     * **Note**: Even if this returns false (search failed), the internal state's mapping will contain
-     * the best partial mapping found, which will be saved by MappingValidator for debugging.
-     *
-     * @param graph_data Indexed graph data
-     * @param constraint_data Indexed constraint data
-     * @param constraints Mapping constraints (for pre-assignment)
-     * @param validation_mode Connection validation mode
-     * @return true if complete valid mapping found, false otherwise (but state still has best found)
-     */
-    bool search(
-        const GraphIndexData<TargetNode, GlobalNode>& graph_data,
-        const ConstraintIndexData<TargetNode, GlobalNode>& constraint_data,
-        const MappingConstraints<TargetNode, GlobalNode>& constraints,
-        ConnectionValidationMode validation_mode,
-        bool quiet_mode = false);
-
-    /**
-     * @brief Get the current search state
-     *
-     * @return const reference to the internal search state
-     */
-    const SearchState& get_state() const { return state_; }
-
-private:
-    SearchState state_;  // Internal state for the search
-    /**
-     * @brief Hash state for memoization (FNV-1a hash)
-     *
-     * @param mapping Current partial mapping
-     * @return Hash value for the state
-     */
-    uint64_t hash_state(const std::vector<int>& mapping) const;
-
-    /**
-     * @brief Recursive DFS search
-     *
-     * @param pos Current position (number of assigned nodes)
-     * @param graph_data Indexed graph data
-     * @param constraint_data Indexed constraint data
-     * @param validation_mode Connection validation mode
-     * @return true if mapping found, false otherwise
-     */
-    bool dfs_recursive(
-        size_t pos,
-        const GraphIndexData<TargetNode, GlobalNode>& graph_data,
-        const ConstraintIndexData<TargetNode, GlobalNode>& constraint_data,
-        ConnectionValidationMode validation_mode);
-};
-
-/**
- * @brief Validates mappings and builds MappingResult
- *
- * Validates complete mappings, checks connection counts according to validation mode,
- * and builds MappingResult with detailed error messages and warnings.
- */
-template <typename TargetNode, typename GlobalNode>
-struct MappingValidator {
-    /**
-     * @brief Validate a complete mapping
-     *
-     * Checks that all edges exist and validates connection counts according to validation_mode.
-     * In STRICT mode: fails if channel counts insufficient.
-     * In RELAXED mode: collects warnings for insufficient channel counts but doesn't fail.
-     * In NONE mode: skips channel count validation.
-     *
-     * @param mapping Complete mapping (mapping[i] = global_idx)
-     * @param graph_data Indexed graph data
-     * @param validation_mode Connection validation mode
-     * @param warnings Optional vector to collect warnings
-     * @return true if mapping is valid, false otherwise
-     */
-    static bool validate_mapping(
-        const std::vector<int>& mapping,
-        const GraphIndexData<TargetNode, GlobalNode>& graph_data,
-        ConnectionValidationMode validation_mode,
-        std::vector<std::string>* warnings = nullptr,
-        bool quiet_mode = false);
-
-    /**
-     * @brief Validate connection counts for all edges
-     *
-     * Checks channel counts and collects detailed warnings/errors according to validation_mode.
-     *
-     * @param mapping Complete mapping
-     * @param graph_data Indexed graph data
-     * @param validation_mode Connection validation mode
-     * @param warnings Vector to collect warnings (must not be nullptr)
-     * @param quiet_mode If true, log errors at debug level instead of error level
-     */
-    static void validate_connection_counts(
-        const std::vector<int>& mapping,
-        const GraphIndexData<TargetNode, GlobalNode>& graph_data,
-        ConnectionValidationMode validation_mode,
-        std::vector<std::string>* warnings,
-        bool quiet_mode = false);
-
-    /**
-     * @brief Print mapping for debugging
-     *
-     * Prints the current mapping showing which target nodes map to which global nodes.
-     * Useful for debugging mapping failures.
-     *
-     * @param mapping Mapping vector (mapping[i] = global_idx or -1)
-     * @param graph_data Indexed graph data
-     */
-    static void print_mapping(
-        const std::vector<int>& mapping, const GraphIndexData<TargetNode, GlobalNode>& graph_data);
-
-    /**
-     * @brief Build MappingResult from search state and mapping
-     *
-     * Converts internal mapping representation to MappingResult, computes constraint
-     * statistics, and collects warnings. Provides detailed error messages when validation fails.
-     *
-     * **Important**: Even if validation fails, the closest/best mapping found is still saved
-     * in the result. This allows users to see what progress was made and diagnose issues.
-     * The `success` field indicates whether the mapping is valid, but `target_to_global`
-     * and `global_to_target` will always contain the best mapping found (if any).
-     *
-     * @param mapping Complete mapping (may be incomplete if search failed)
-     * @param graph_data Indexed graph data
-     * @param state Search state (for statistics and error messages)
-     * @param constraints Mapping constraints (for computing constraint stats)
-     * @param validation_mode Connection validation mode
-     * @return MappingResult with success status, mappings (always saved even if invalid), warnings, and statistics
-     */
-    static MappingResult<TargetNode, GlobalNode> build_result(
-        const std::vector<int>& mapping,
-        const GraphIndexData<TargetNode, GlobalNode>& graph_data,
-        const DFSSearchEngine<TargetNode, GlobalNode>::SearchState& state,
-        const MappingConstraints<TargetNode, GlobalNode>& constraints,
-        ConnectionValidationMode validation_mode,
-        bool quiet_mode = false);
-};
 
 }  // namespace tt::tt_fabric::detail
